Add files via upload
Hey there. This project is a machine learning model; the model, a decision tree, utilizes features such as pH, hardness, sulfates, among others, to estimate whether a water sample meets potability standards.

The methodology and implementation of the model are inspired by various existing literature sources, adapting them to create a system that fits a variety of use cases. The code's flexibility will allow other researchers and professionals to use parts of the code or the entire code for their own data sets and specific circumstances.